move access token check to query
